Added pip and setuptools as nativeBuldInputs to allow check.
I had to set doCheck to false before in order to get the package to
build, otherwise it would complain that it couldn't find pip.
Adding pip and setuptools as nativeBuildInputs fixes this and allows
for doCheck to be true. This will be useful for debugging the build.
Now that the check is working it complains that it can't find the
correct version of uvicorn.